% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{corner.label.w}
\alias{corner.label.w}
\title{corner.label.w}
\usage{
corner.label.w(label = "A", cex = 3, x = -1, y = 1, xoff = 1,
  yoff = 1, figcorner = T, ...)
}
\arguments{
\item{label}{Text to display}

\item{cex}{font size}

\item{x}{an integer value: -1 for the left side of the plot, 1 for the right side}

\item{y}{an integer value: -1 for the bottom side of the plot, 1 for the top side}

\item{figcorner}{Whether to find/display at the corner of the plot or figure.}

\item{`xoff, yoff`}{Horizontal and vertical text offsets. If NA, it defaults to one half of the width and height of "m" respectively.}
}
\description{
Add Legends to the corners. From the Plotrix package.
}
\examples{
corner.label.w("A")
}
