% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{md.LinkTable}
\alias{md.LinkTable}
\title{md.LinkTable}
\usage{
md.LinkTable(tableOfLinkswRownames)
}
\arguments{
\item{tableOfLinkswRownames}{a dataframe where every entry is a string containing an html link}
}
\description{
Take a dataframe where every entry is a string containing an html link, parse and write out a properly formatted markdown table
}
\examples{
tableOfLinkswRownames(tableOfLinkswRownames = df_of_LinksParsedByDatabaseLinkeR)
}
