% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{stopif}
\alias{stopif}
\title{stopif}
\usage{
stopif(condition, message = "")
}
\arguments{
\item{condition}{any condition check that gives TRUE or FALSE}

\item{message}{print a message}
}
\description{
Stop script if the condition is met, and print a message
}
\examples{
a=1; stopif (a==1, message ="A is 1")
}
