% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wvenn}
\alias{wvenn}
\title{wvenn}
\usage{
wvenn(yourlist, imagetype = "png", alpha = 0.5, fill = 1:length(yourlist),
  subt, ..., w = UnlessSpec("b.defSize", 7), h = w,
  mdlink = ww.set.mdlink(), plotname = substitute(yourlist))
}
\arguments{
\item{yourlist}{The variable to plot.}

\item{imagetype}{Image format, png by default.}

\item{alpha}{Transparency, .5 by default.}

\item{fill}{Background color vec}

\item{subt}{Subtitle}

\item{...}{Pass any other parameter of the corresponding venn.diagram() function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{plotname}{Manual plotname parameter}
}
\description{
Save venn diagrams. Unlike other ~vplot funcitons, this saves directly into a .png, and it does not use the dev.copy2pdf() function.
}
\examples{
wvenn (yourlist =  , imagetype = png, alpha = 0.5, ... =  , w = 7, h = w, mdlink = F)
}
